IO: fixes variable number of comment lines
In the ascii data files, the number at the beginning of the file tells
us how many comment lines follow. This number instead of the hard-coded
values is then used to skip over the comments. This is not really save
and should be changed to check for '#' at the beginning of a line.