Messung_Nadir: vectors instead of arrays
Uses vectors instead of the dynamically allocated arrays for the stuff
in Messung_Nadir. This is analogous to the things in Messung_Limb and
we also introduce a helper function for the IO stuff dealing with the
initialisation from file date.

As a side effect, we don't need to allocate and de-allocate the memory
and we can get rid of the destructor.