IO: bugfix, intensity correction
In the original code for larger heights, the intensity is corrected by
subtracting the radiance measured at the largest height.
Since we are at it, we also use the index given by no_of_alt which might
not always be 31.

Last, we set he offset in ReadL1C_Limb_meso_thermo_mpl_binary_reduziert()
to Anzahl_Hoehen-1 to prevent a possible reading outside the vector
boundaries (not perfect, yet).