main: use config retrieval covariances
Uses the config values for the retrieval covariances by setting an
additional index for the species: 0 = MgII, 1 = MgI.

There is no need to re-compile the code and the Lambdas are printed to
stdout, making it possible to re-run the calculations, provided stdout
is saved along with the rest of the output, e.g. via tee(1).