IO: simple straylight detection in the dark signal
This includes a simple logic to distinguish a dark signal with
straylight structure from a flat dark signal. If straylight is
detected, the old point-wise subtraction of the spectra is used.
Otherwise, the average or median of the dark signal is subtracted.

Straylight is detected, if the averaged dark signal is larger than a
specific cut-off, which is hard-coded to be 6.e-9 for now.