sciamachy convolution blur
Introduces a sciamachy blur function intended to be used for
high-resolution solar spectra from, e.g. SAO.
The weight vector is quite large since the resolution differs by an
order of magnitude (0.01 nm reference vs. 0.11 nm SCIAMACHY).

To be useful, the blurred spectrum has to be downsampled to the
SCIAMACHY bin-width.