limb: adjust geo-locations after pole
The tangent point (TP) latitude is increasing again in the south after
the TP has passed the pole. This might lead to mis-ordering the TPs
within an orbit.

This adjusts the geo-locations of TPs after the pole by extending theta
to be smaller than -90 deg and by flipping phi around. Now, these points
should all come after the last regular limb TP of the orbit and should
be treated correctly by the retrieval.