NO: coefficient printing methods
To avoid exposing the internal data, here are some printing routines to
check whether we have read and calculated sensible data and to be able
to compare with the IDL program. Thus, the main purpose is debugging.

Also adds some functions to expose the emission lines and gammas to the
outside since that's what is really needed there.