postpone NO wavelength setting
Postpones the setting of the NO wavelength (in our case the one of the
maximum emission) to the limb analysis and sets most of the species data
to some dummy values. This makes sure that the matrix algorithm find the
right E1 and E2 for NO.

Also adds a missing clean-up before the NO species data is initialised.