limb analysis: recalc (almost) everything in the NO case
The emissivities in the NO case depend on the temperature which, in
turn, depends on the actual tangent point position and time. Thus, we
create a new object within the limb analysis loop which uses only the
transition parameters from the given NO_emiss reference.
All the rest (emissivity calculation, SCIA resolution convolution, etc.)
is done again.