NO limb: public matrix 'solar'
Makes the solar index matrix public for easy copying. That way, we don't
need to read it in again in Limb_Auswertung() and a different one than
given by Solspec can be used.

Solspec is intended to be the SCIA solar spectrum which is unusable for
the NO emission calculation. We need a different reference for that.