air mass matrix: bugfix in pixel finder
Fixes a min/max typo in the pixel finder.

The grid is sorted from northern to southern latitudes (as are the
measurements) which means that the largest (= northernmost) latitude is
Grid.m_Gitter[0].m_Max_Breite and the smallest (= southernmost) is given
by Grid.m_Gitter[<gridsize> - 1].m_Min_Breite. This min/max was mixed up
when testing if the point is within the bounds.

Also changes the prepared (and commented) debug output accordingly.