config: vector with NO transition data
Instead of the data for a single (rotational) transition, more NO
transitions can be defined in the config file and the upper and lower
levels are stored in a std::vector.

Adapts the NO species preparation in main() to use the config vector
data by looping over all configured transitions.