matrices: defer S_y generation to after raytracing
Introduces an extra function to generate the measurement covariance
matrix S_y from the measurement fit errors and defers this generation to
after the ratracing step. This allows us to change the numbers in the
vectors with the analysed measurements during raytracing, e.g. the
slant column densities and their errors.