NO: fix solar flux from ph/s/cm^2/nm to ph/s/cm^2/Å
The solar radiance read from the file is given in ph/s/cm^2/nm whereas
the NO emissivity calculation needs it in ph/s/cm^2/Å which translates
to a factor 1./10. for the solar radiance.

This increases the slant column densities by about this factor and the
retrieved number densities by about a factor 3 which brings them in line
with the results from other experiments, e.g. MIPAS and SNOE.