limb analysis: use wavelet approximated spectra
Uses wavelet approximation of the spectra for the fit of the NO line(s).

This involves two runs of the fit routine to first extract the wavelet
coefficients for each limb point. Then the median (or average) of the
last and second to last components (of the components of the whole scan)
is used to approximate the spectrum for each limb point for the NO line fit
and the actual fit is done.