limb: fixed offsets for NO fit window
Uses fixed indices instead of a search through the NO line to find the
minimum and maximum wavelength of the fit window. This is required for
the wavelet analysis and reconstruction/approximation step since all the
limb spectral windows need to have the same length.

The starting point right now is the wavelength of the maximum NO
emission.