limb NO: always interpolate large peaks
Interpolates large peaks in the spectra by setting the respective
radiances directly. This is done when we build up the fit spectrum for
the solar fit and the interpolated radiances are used by everything that
follows when fitting the NO spectra.

The large spikes are interpolated by smoothing three points, the one
before and the one after them are also interpolated.