validity tests: fail SAA if the spectrum is unusable
Requires that the number of spectral points for the SAA check is not
zero, otherwise it returns true (= SAA) since the spectrum is unusable
for the SAA check and probably other things too. This skips this one for
the retrieval and shifts the real SAA detection to use another spectrum
if available.