skip matrix memory allocation at initialisation
Removes the sizes in most of the MPL_Matrix initialisations.  The memory
allocation of them is done in Matrizen_Aufbauen() so that the previously
allocated memory was never used and only increased the memory footprint.

Fixes the allocation of the s priori covariance matrix, and as a result,
only memory for the retrieved species is allocated.