NO: reorganise the scia convolution
Reorganises the SCIA convolution by using member variables and doing
the iteration over the limb spectrum only up to the second to last
element. This saves us the check for xdiff every time, and it is
further out than we will ever need, anyway.
Eliminates the call to std::distance for the running index, saving a
few cycles, and sets the last element of the convolved spectrum to zero
for safety.