MPL_Matrix: fix shadowed methods
Renames Zeile() and Spalte() to get_...() to avoid them being shadowed
by the local initialisation variables in some other methods, in
particular in the operator() declarations. Since these methods are
not yet used anywhere in the code, renaming them results in a smaller
change than doing it the other way around, and it even makes their
purpose clearer.