IO: interpolate the 282.03 nm pixel
The pixel at 282.03 nm is at the border of the two subchannels with
sometimes different integration times. This leads to a unusable peak in
the spectrum and we interpolate over it.
We do this by saving the index and using this value later since
282.03 nm may be at different positions depending on the scan mode,
e.g., MLT, nominal, or nadir scans.