SAA test: change from max to peak-to-peak value
Changes the SAA check to use the peak-to-peak intensity instead of just
the maximum value. This helps detecting unusual spikes in the spectra
when the dark limb spectrum has already been subtracted at calibration
time. This happens when using the SCIAMACHY nominal states that are
calibrated with the '-darkflag LIMB' setting, which is the default.