limb: read all nominal spectra
Sets Anzahl_Hoehen to -1 to read the spectra from all tangent altitudes
of a nominal mode scan. The usable ones (from 50 to 200 km) are selected
later for the measurement vector. The overall number of limb points
varied with time and this makes sure that we use all available spectra
in this altitude range for the retrieval.