matrix preparation: catch "mysterious" cases
Includes the lowest grid altitude in the last check in
Pixel_finden_und_AMF_erhoehen_LOS() to prevent "mysterious" errors if
Punkt_Hoehe == Grid.m_Gitter[0].m_Min_Hoehe.
The search through all grid point afterwards excludes the lowest
altitude of the grid point and the check for equality was missing.

Note: Sometimes the calculations in Luftmassenfaktoren_Matrix_aufbauen()
lead to rounding errors in Punkt_Hoehe, making it slightly smaller than
<grid_point>.m_Min_Hoehe. The correct way might be to add a small
tolerance to the exhaustive search, but we leave it like this for now to
not break compatibility with previous calculations.