retrieval error: unnecessary transpose in matrix inversion
We don't need to transpose the matrices for the correct memory layout.
The inverse is calculated using A * X = 1, but using C-style (row-major)
memory layout, the Fortran code solves A^T * X = 1.
The result is X = (A^T)^-1 = (A^-1)^T, such that X^T = A^-1.
Therefore, we can also skip the second transposition because the output
of dgesv_() already is transposed in C-style memory layout.