MPL_Matrix: introduce cheap transposition
Introduces cheap matrix transposition using a 'transposed' flag to
determine the storage layout in memory.
The relevant use cases are the (i, j) access operator and the matrix
multiplication operator*. Updates both to promote the correct layout.
In the second case to lapack's dgemm_() using the correct
'transa', 'transb', 'lda', and 'ldb' calues.

Renames the old method to transponiert_full() and introduces an
alternative version transponiert_full2() for the rare cases when we
need an actual re-arrangement of the memory layout.

Uses the full transposition methods before calling dgesv().