retrieval: re-order decomposition and initial solution
Re-orders the LU composition and initial solution for the density
retrieval. We now solve and decompose at the same time using dgesv_()
instead of first decomposing and solving a dummy equation using dgesv_()
and then using dgetrs_() for the initial solution of the retrieval
equation. This saves only a few cycles by skipping dgetrs_() but the
code is now a little more structured and straight-forward.