NO regression model python code
This is a simple NO model based on the multi-linear regression
coefficients from the MLT data from ACE-FTS, MIPAS, SCIAMACHY, and SMR.

This is the python version and includes the needed data files for
successful applying the model. Most of the needed variables are set via
a json file. The whole code resides in the DATA/ directory within the
main source tree. It needs to be copied or linked for using it.