Use the common measurement class
Uses the common measurement class and derives the limb and nadir
classes from it. This remove lots of duplicate code and makes
maintenance easier.

We now only keep variables and methods differing between limb and nadir.

Corrects the call to Intensitaeten_durch_piF_Gamma_berechnen() in
Nadirauswertung.cpp for the common measurements method.