nadir analysis: remove separate measurement number
Removes the additional argument for the measurement number and
replaces its use by the member variable m_Messung_ID in
Messung_Nadir::Zeilendichte_Bestimmen().

It is only used for the filename for the spectral fit plot and both are
equivalent. Hence, we don't need to carry it over by an additional
function argument. This brings us also more in line with the limb case.