limb/nadir: Move member variables to general class
Moves the special limb and nadir member variables to the general
measurement class. This simplifies and avoids duplicated code.

All members are now initialised in Messung::Messung(). We simply
use this in Messung_Limb::Messung_Limb() and
Messung_Nadir::Messung_Nadir(). We also don't need the overloaded
assignment operators anymore.

Makes it easier to use the same methods for spectral fits later on.
This also reduces duplicated code and the risk to introduce errors.