IO: Safeguard header reading from binary files
In the usual case, we forcefully read a number of lines (equal to fixed
100 char chunks) from the binary spectra file. We then suppose that the
"real" data starts right after N 100 byte chunks. This fails in the
unlikely case that the real text header is shorter and the real data
starts earlier in the file.

We solve this by looking for a hash (#) sign at the beginning of the
100 byte chunk. If it is different, we jump back to the saved position
and abort the text  header loop. Then, hopefully, we continue to read
the data as usual. We do this both for limb and nadir files.