fit plots: Update python figure layout and style
Changes the line colours to use the purple/orange mix from colorbrewer2
to be colourblind and greyscale safe.

Includes "$" in the axis labels to render math content correctly.

Shifts the numbers to the upper left to not hide the data lines as it
was sometimes the case.

Adjusts axis ticks and labels to be consistent with all our other plots.