Clarify that S_y and S_apriori are S_y^-1 and S_a^-1
If we compare the retrieval equations used here to the standard
references (Rogers for example), S_y in our code is actually S_y^-1
in those references. The same applies to S_apriori (~ S_a^-1).

Therefore, we add a comment to the relevant code such that any user
reading the code knows it and gets a little bit less confused.