matrices: Flexible a priori transition function
Renames the a priori 'scale_function_'s to 'transition_' since
this describes their purpose best. Then replaces the hard-coded
transition (scale_function_smooth()) by a function pointer which
may be changed at runtime. This allows to use a wider range of
transition functions, for example differentiable functions
from 0 to 1 that are continuous and differentiable at the end points.