nadir analysis: Include absorption wavelength
Includes the absorption wavelength in the analysed nadir scans.
Sets the value accordingly in the case of NO.

In the other cases (matals), the absorption wavelength is equal to the
emission wavelength (resonant emissions). Switches from using the
species variables to using the line parameter variables. One more step
to make the limb and nadir case consistent.