MPL_Matrix: Configurable precision in matrix output
Makes the matrix output precision configurable. The default should
usually be good enough but in the case of debugging, a higher precision
may be needed. Introduces an optional second argument to
MPL_Matrix::in_Datei_speichern() which gets honoured if it is set to
anything different from zero. Otherwise the output precision is left
alone to be compatible with previous code versions.