CMake build file
Adds CMakeLists to aid building the executable.
May not be perfect but it works and should be good enough for now.
Includes all needed libraries and includes, works at least on x86
Solaris with sunperf and openblas.
Optional things are commented and may later be used for debug builds
or similar.