Honour configured retrieval grid extensions
The configuration file has the possibility to extend the retrieval
grid above the top grid altitude. This includes these points by
extending the grid vertically, making sure that in all relevant cases
that occurred so far, the old behaviour (no extension whatsoever) is
retained. We achieve this by requiring that the lowest extension
altitude must be above the top grid altitude. Otherwise we will not
extend the grid.