Retrieval: Fix calculating x_1
According to [Funke 2005, Steck & von Clarmann 2001] (and probably
Rodgers), the iterative solution is
x_{i+1} = x_i + (K^T S_y^{-1} K + R)^{-1}
            \times [K^T S_y^{-1} (y - F(x_i)) + R (x_a - x_i)].
We can choose x_0 arbitrarily, and in the case of x_0 = x_a,
(x_a - x_0) vanishes, but we have to add x_a to the first solution.
Apparently here we use x_0 = 0 and since F is linear (F(x) = K*x in our
case), F(0) = K*0 = 0. Therefore, y - F(x_0) = y for calculating x_1.

This corrects this in the code and adds the above comment to the
appropriate place. First tests suggest that this does not change the
results at all, which seems plausible because the removed term will
be correctly taken into account in the first iteration when
calculating x_2. Furthermore, this change only affects the logic
when x_a != 0.