meas: Separate Rayleigh fit with peak interpolation
Separates the Rayleigh fit (with peak interpolation) from the main
NO spectral fit function. The fit factor is used to subtract the
Rayleigh background from the limb spectrum.

We split it and make it more flexible such that, e.g., we can alter the
wavelength range of the Rayleigh fit. The idea is to allow using the same
Rayleigh background for all NO transitions.

Does not change the NO results.