smooth: Allow matrices as rhs in my_solve()
Replaces the fixed number of columns (set to 1) by the actual number of
columns of the right-hand-side matrix in my_solve(). As long as it has
only one column, nothing should change. But we can now solve more
equations simultaneously, in particular we prepare to invert matrices
with it, using the identity matrix as rhs input.