IO: Implement reading spectra from netcdf files
Implements a new function to read the spectra from our self-made
netcdf files, converted using our python code. This involves a lot
of manually "transposing" the data, because the attributes are stored in
an array for all altitudes. However, the following code expects all of
them grouped together per altitude in a "Limb_Datensatz".