main: Implement adaptive NO regularisation
Using the calculated fit variances, we estimate the lambdas used to
regularise the retrieval solution. Our empirical formula matches the
"standard" lambdas for the usual cases but should better balance the
regularisation and the fit errors if these deviate too much from the
"standard" case.

In particular, if we use the correct error calculation, the fit errors
are smaller (about half) and we have to adjust the lambdas manually.
Using the adaptive method, we get approximately the same retrieval
result in both cases.

The adaptive method is triggered by using a negative a priori lambda.
This allows easy switching between the two variants.