plots: Modify and beautify labels
Uses regular expressions (std::regex_replace()) to prepare the labels
in the python script for a nicer look. In the numbers we replace the
order of magnitude ("...e+mag") by \times 10^{mag}. We also set only
the unit in math mode (as "$[unit]$") instead of the whole x or y label.

Additionally, we replace all white spaces by small latex spaces ("\,")
in the axis units. This gives them a nicer look since otherwise the
math mode removes the spaces.