MPL_Matrix: Implement writing to netcdf
Instead of using the gzipped text, we can now write the matrices to
(compressed) netcdf-4 files. We use single precision floats, which saves
a few bytes of disk space and which can be further reduced using the 32
bit (unsigned) integer representation of the data as described in
http://nco.sourceforge.net/nco.html#Packed-data .

It includes primitive error checking by simply returning early and
elevating the error code.