MPL_Matrix: Implement saving as hdf5
In addition to gzipped text and netcdf4, we can now write the matrices
to (compressed) hdf5 files. As in the case of netcdf4, we use single
precision floats which can be further reduced using the 32 bit
(unsigned) integer representation of the data as described in
http://nco.sourceforge.net/nco.html#Packed-data .

Hdf5 does not need dimensions so we can save the data as-is. However,
we have to define chunk sizes manually for the compression to work.

It also includes primitive error checking by simply returning early and
elevating the error code.