IO: Fix relative radiance error calculation
We calculated the error assuming we deal with absolute errors.
This was not correct, the errors given are relative errors.
They should be treated as such when subtracting the dark signal.
We fix this by converting them to absolute errors (squared), and then
dividing the result by the new, dark-subtracted radiance value.

We use the opportunity to make the whole block more readable by using
temporary variables for the signal and the relative error. It should be
much clearer and easier to understand this way.