config: Add switch for NO polarisation correction
Since the effect of this correction is at least questionable, we
introduce a configuration option to easily switch it on or off.
The default is on for compatibility reasons.

Changes the default config files to include this switch.

Also fixes an order complaint of the compiler regarding
m_min_TP and m_max_TP.