Fix solar spectra reading of D0 files
Solar files named "D0" contain more header lines and different meta data
compared to the files without "D0" in their names.

We distinguish between them by looking at the first line of the file.
If it starts with a '#', we assume that it is an ordinary file.
In the other case, we use the number as the number of header lines
before the solar spectral data.