main: More robust high res solar spectra detection
Previously we relied on the overall number of wavelength points being
very large to distinguish highly resolved solar spectra and SCIA solar
spectra. We now look at the spacing using std::adjacent_difference()
and its median and convolve everything that has about twice the SCIA
resolution.

We also add a short notice when we convolve the spectra for later
reference and debugging.