main: Implement finding the orbit solar spectrum
Introduces a function to find the solar spectrum file for the current
orbit by looking at the first line of the orbit file list. From this
we extract the directory in which the limb (or nadir) files reside and
then use glob(3c) to find the corresponding solar spectrum file.