main: Determine solar spectrum from orbit list
Instead of using a fixed solar spectrum and scaling, we try to find the
corresponding orbit solar spectrum from the filenames in the orbit list.
This should work because the list contains absolute paths and the solar
spectrum resides usually in the same directory. However, to keep
backward compatibility, this is only done if the configured solar path
is set to "auto". Otherwise the given file will be used.