limb: Determine solar factor from reference
The high resolution solar reference spectrum (Chance and Kurucz, 2010)
is normalised to 1 AU. To estimate the intensity for the NO gamma band
excitation, we fit the SCIA measured solar spectrum to the reference.
These factors were formerly prepared beforehand and saved in a file
for each day. This adds the possibility to calculate this factor
on-line from the given (or determined) SCIA solar spectrum.
It also gives more fine-grained factors since it is done for each orbit
instead of daily.

To keep consistency with previous code versions, the automatic method
only kicks in if the solar correction factors filename is set to "auto"
in the configuration file.