Update configuration templates with solar changes
Updates the configuration templates with useful values to keep track of
the recent code changes. Note that currently the solar path has to be
set on the command line so that setting it here is a no-op. But we
change it to reflect that "auto" should be the new default.