limb: Use prepared Rayleigh class for fit
Uses the above declared class rayl{} to determine the Rayleigh fit
factor in Messung::fit_rayleigh_and_interp_peaks().
This assures consistency with the background subtraction later in the
code in Messung::slant_column_NO(). If we want to do these things more
sophisticated, we have to change things only in one place.