Fix configuration templates [1/2]
We actually care now about the number of grid extensions above the
nominal grid altitude. Since this is mainly unused during the normal
retrieval, we set them to zero.

Also fixes the absorption cross sections, we already updated and used
the new ones in the regular runs. We here make the options consistent
for reference.