limb: Include F10.7A value in MSIS inputs
Previously we used the same values for the F10.7 and F10.7A input
parameters of the MSIS model for the temperature calculation.
Since we now also have the F10.7A (81-day centered moving average)
values available, we use them for the MSIS model to run a little bit
more exact. We also add the used F10.7A value in the debug output
(should go to some logfile).