Change altitudes for NLC (PMC) detection
Increases the starting altitude of the monotony check to really cover
all NLC altitudes (84 km to 93 km). Previously this started lower and
therefore failed to detect NLCs at higher altitudes.

We also remove the temporary return variable and just return early if we
detect a NLC. This may save a few cycles.