NOEM: Allow "automatic" scaling based on F10.7
Solar activity was very high during the SNOE measurement period used for
the NOEM model. Until now, we accounted for this by suggesting a constant
scaling of 0.5 for the a priori. To counter further criticism and to
account for different solar activity during the SCIAMACHY period, we
allow for empirically scaling the NOEM a priori by F10.7 / 150.
150 is about the mean F10.7 during the SNOE time period used for NOEM.
This "automatic" scaling is enabled by setting a negative scale factor
in the config file.