grid: More sophisticated longitude interpolation
Instead of interpolating the measurement longitudes to the equator,
we calculate the theoretical crossing latitude for each limb scan.
This we do using the reverse relation to lon = C * tg(lat) + lon0.
We then take the equator crossing longitude as the mean of all
calculated crossings (except the orbit endpoints).
To avoid periodicity problems, we again use sine, cosine and atan2().