retrieval: Fit a priori scale
Adding a free parameter to allow fitting a scale factor for the a priori
gives a coupled system. We resolve this by substituting one into the
other and then calculating the fit factor from the initial retrieval
run. This seems to work equally well when compared to extending the
matrices and iteratively fitting the scale factor.

The a priori scaling fit is activated by setting the fit factor to -2 in
the configuration.