main: Implement NO a priori and covariance scaling
Implements the NO a priori scaling determined by the ratio of the
slant column densities from one forward run with the a priori to the
measured ones (option -3). This factor is calculated in more cases to
have a value to compare to. But the prior is only scaled if requested,
i.e. if Konf.NO_apriori_scale is -3.

The prior covariance can be scaled differently from the standard setting
in the case the a priori it is not zero. The relevant configuration
options are Konf.NO_apriori_cov_relative to scale it relative to the a
priori value (e.g. 200% or 500% or something) or not.
Konf.NO_apriori_cov_factor adjusts the (weight) factor. In the case of
relative scaling we use S_a = 1 / (lambda^-1 + (f * x_a)^2)), in the
absolute case (Konf.NO_apriori_cov_relative = 0) we use S_a = lambda / f.