NO regress model: Add FIXME comment
Because we just copied the code from the original regression model,
some variables are not defined. They have to come from outside which
is unpractical without further knowledge.

We add a big FIXME to the part of the code which will be removed
in the future. It cannot be used as it is now.