Include local copy of gzstream
Until now we linked to an externally compiled gzstream lib, but since it
is LGPL, we may as well distribute it (we already included the header).
This makes it easier to compile the code from scratch.

Adapts the cmake file, e.g. gzstream.cpp needs to find its header for
compiling.

Also includes the gzstream COPYING file (LGPL).