cmake: Clean up
The Polarization_correction files are not part of the code and should
not be there in a fresh clone. So we don't need to filter them.

Removes the architecture flag (-m64) for more consistent 32/64 builds.
The compiler should know what to do and one can always use
CFLAGS/CXXFLAGS/LDFLAGS if those need to be specified explicitly.