README: Show how to create the orbitlist file
Adds the simple find command that can be used to produce the orbitlist
file that we need as an input. This should help to understand how to
run the code.

Wraps the command line arguments in <> as it is commonly done.
Also links to the (new) orbitlist subsection for details.