cmake: Improve library searching
Enables to specify the library search paths for openblas, sunperf, and
NOEM on the command line to find them in non-standard locations. The
paths can be set by defining OPENBLAS_LIBS, SUNPERF_LIBS, NOEM_LIBS,
or by listing them in CMAKE_LIBRARY_PATH. Note that the paths in
CMAKE_LIBRARY_PATH are separated by semicolons which may have to be
escaped depending on the shell.

Examples:

    $ cmake -DOPENBLAS_LIBS=/path/to/openblas∕lib -DNOEM_LIBS=/path/to/NOEM/lib /path/to/src
    $ cmake -DCMAKE_LIBRARY_PATH=/path/to/openblas∕lib\;/path/to/NOEM/lib /path/to/src

(Removes the last references to hard-coded "local" installation paths.)