matrix mul: Avoid explicit transposition
Before calling LAPACK's dgemm_() for matrix multiplication, we already
make sure that we pass the correct memory layout based on the transposed
flag. The result will be in Fortran order which means transposed to C.
Instead of explicitly copying the entries back to C order, we just set
the transposed flag.

The places where it really matters is element access and further matrix
multiplications. Both methods take care of the transposed flag.

Directly filling the result avoids the temporary matrix which
should save memory and a few cycles for copying the elements.