matrix: Fix output to disk
For consistent disk output with previous versions we fix the in-memory
order of any transposed matrix before writing them to disk.
We use an intermediate matrix such that the output routines stay pure,
changing the transposed methods also to `const` (i.e. pure functions).
Saving to disk usually happens only once at the end when everything else
is finished and the impact of transposing on runtime and memory should
not matter anymore.