retrieval: Update S_a to speed up AKM calculation
We separated the constraints on the magnitude (S_a) from those on the
form (R) of the solution `x`. The averaging kernel is calculated from
the full constraint matrix S_a + R. At the inversion step, we already
have both available and instead of calculating R again to calculate AKM,
we update S_a in place after the inversion. This saves the costly matrix
multiplication for the uncertainty analysis, speeding up the AKM and S_x
calculation.