sol: Better logic for 2/3 column spectral files
We have different versions of the solar spectrum, depending on how the
calibration was carried out. We used to discriminate based on the number
of header lines, but there is apparently a version that has the same
header but three instead of two columns. Fortunately, the header
mentions the additional column "accuracy", so we check for the existence
of it to chose between two and three column solar spectrum files.