main: Fix solar spectral filename logic
Previously we relied on the limb state filename to contain a ".dat"
before the final ending (such as ".l_mpl_binary") to extract the orbit
number. This ".dat." is not necessarily the case, and the pathname may
also contain dots. Therefore, instead of looking for the last dot,
we extract the filename first and then look for the first dot, which
should appear consistently right after the orbit number in the filename.